# Conflicts:
#	app/src/main/java/se/kth/csc/iprog/dinnerplanner/model/DinnerModel.java

+++ app/src/main/java/se/kth/csc/iprog/dinnerplanner/model/IDinnerModel.java
@@ -25,7 +25,7 @@ public interface IDinnerModel {
 	/**
 	 * Returns the total price of the menu (all the ingredients multiplied by number of guests).
 	 */
-	public float getTotalMenuPrice();
+	public double getTotalMenuPrice();
 
     /**
      * Adds the passed dish to the menu. If the dish of that type already exists on the menu
